feat: new file upload component with drop
