+++ src/Template/Element/Form/relations.scss
@@ -7,29 +7,14 @@
             }
         }
     }
-
-    .drop-area {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        width: 100%;
-        min-height: $gutter * 6;
-        color: $gray-600;
-        background-color: $gray-950;
-        transition: .2s transform;
-        &:after {
-            content: attr(placeholder);
-        }
-        &.dragover {
-            transform: scale(1.05);
-            color: #fff;
-            background-color: #000;
-        }
-    }
 }
 
 // related item (relate-item-twig)
 .related-item-column {
+    &.sorting {
+        opacity: 0.2;
+    }
+
     article {
         height: 100%;
 
@@ -83,6 +68,83 @@
     &.removed {
         article {
             opacity: .2;
+        }
+    }
+}
+
+
+// drop and upload
+.drop-area {
+    position: relative;
+    width: 100%;
+    min-height: $gutter * 5;
+    background-color: $gray-950;
+    transition: .2s transform;
+
+    &.dragover {
+        transform: scale(1.05);
+        color: #fff;
+        background-color: #000;
+    }
+
+    .upload-placeholder {
+        color: $gray-550;
+        margin: auto;
+    }
+
+    .file-input {
+        @extend .full-size-absolute;
+        opacity: 0;
+        outline: 0;
+        cursor: pointer;
+    }
+
+    .upload-items {
+        .upload-item {
+            width: 100%;
+            & ~ .upload-item {
+                margin-top: $gutter;
+            }
+
+            .upload-item-header {
+                display: flex;
+                justify-content: flex-end;
+                align-items: flex-end;
+                .name {
+                    margin-right: auto;
+                    word-break: break-word;
+                }
+                button {
+                    margin-left: $gutter * .5;
+                }
+            }
+
+            .message {
+                line-height: 1.25rem;
+                min-height: 1.25rem;
+            }
+
+            .progress-bar {
+                width: 100%;
+                margin: $gutter * .75 0 $gutter * .75;
+
+                .progress-bar-status {
+                    min-width: $gutter;
+                    height: 2px;
+                }
+                .progress-bar-status.in-progress {
+                    background-color: $gray-400;
+                }
+                .progress-bar-status.done {
+                    background-color: $info;
+                }
+                .progress-bar-status.error {
+                    background-color: $error;
+                }
+                .progress-bar-status.cancelled {
+                    background-color: $gray-700;
+                }
+            }
         }
     }
 }
@@ -151,7 +213,7 @@
     }
 
     // ui specs
-    .related-object article {
+    .related-item-column article {
         cursor: pointer;
 
         &.selected {
@@ -162,13 +224,13 @@
                 justify-content: center;
                 align-items: center;
                 position: absolute;
-                width: $gutter * 2; height: $gutter * 2;
+                width: $gutter * 2.25; height: $gutter * 2.25;
                 top: $gutter * .5; left: $gutter * .5;
                 border: 2px solid white;
                 border-radius: 50%;
                 background-color: $info;
                 color: #fff;
-                font-weight: $font-weight-bold;
+                font-weight: $font-weight-black;
             }
         }
     }
