+++ src/Template/Element/Panel/panel.scss
@@ -32,51 +32,3 @@
         padding-top: $gutter;
     }
 }
-
-// multiple file upload .../app/components/upload-files-view.js
-.main-panel-container .upload-files {
-    .upload-item {
-        & ~ .upload-item {
-            margin-top: $gutter;
-        }
-
-        .upload-item-header {
-            display: flex;
-            justify-content: flex-end;
-            align-items: flex-end;
-            .name {
-                margin-right: auto;
-            }
-            button {
-                margin-left: $gutter * .5;
-            }
-        }
-
-        .message {
-            line-height: 1.25rem;
-            min-height: 1.25rem;
-        }
-
-        .progress-bar {
-            width: 100%;
-            margin: $gutter * .75 0 $gutter * .75;
-
-            .progress-bar-status {
-                min-width: $gutter;
-                height: 2px;
-            }
-            .progress-bar-status.in-progress {
-                background-color: $gray-400;
-            }
-            .progress-bar-status.done {
-                background-color: $info;
-            }
-            .progress-bar-status.error {
-                background-color: $danger;
-            }
-            .progress-bar-status.cancelled {
-                background-color: $gray-700;
-            }
-        }
-    }
-}
