+++ src/Template/Element/Panel/panel.twig
@@ -6,6 +6,5 @@
         <relations-add v-if="props.action === 'relations-add'" v-bind="props.data" inline-template>
             {% element 'Panel/relations_add' %}
         </relations-add>
-        <upload-files-view v-if="props.action === 'upload-files'" v-bind="props.data" />
     </div>
 </panel-view>
