+++ src/Template/Layout/styles/_elements.scss
@@ -303,26 +303,3 @@ a.close-button {
     height: 1px;
     margin: 1.5rem 0;
 }
-
-
-// TODO: controllare questo dragover
-// .dragover {
-//     position: relative;
-//     &:after {
-//         content: "Drop Here";
-//         @extend .full-size-absolute;
-//         z-index: 100;
-//         display: flex;
-//         align-items: center;
-//         justify-content: center;
-//         background: rgba(0, 0, 0, .85);
-//         font-size: $font-size-2;
-//         font-weight: $font-weight-bold;
-//     }
-// }
-
-.sorting {
-    background: $black;
-    opacity: 0.4;
-    border-radius: $gutter * 0.25;
-}
