+++ src/Template/Layout/styles/_variables.scss
@@ -23,6 +23,7 @@ $font-weight-light:           200 !default;
 $font-weight-normal:          300 !default;
 $font-weight-medium:          400 !default;
 $font-weight-bold:            500 !default;
+$font-weight-black:           800 !default;
 $font-weight-base:            $font-weight-normal !default;
 
 // line
